hard-code partitions

+++ src/dst/pgsql.py
@@ -76,10 +76,11 @@ class PgSQLSaver:
             ## Drop indexes if exist (except primary key index)
             for index_name, index_def in indexes.items():
                 with conn.cursor() as curs:
-                    sql_stmt = f"DROP INDEX IF EXISTS {index_name};"
-                    self.inform(sql_stmt)
-                    curs.execute(sql_stmt)
-                    conn.commit()
+                    if(index_name != f"{table}_pkey"):
+                        sql_stmt = f"DROP INDEX IF EXISTS {index_name};"
+                        self.inform(sql_stmt)
+                        curs.execute(sql_stmt)
+                        conn.commit()
             self.inform("Dropped indexes")
         conn.close()
         return exists
